consolidating tracking manager and location manager to avoid state issues
