--- Rear Rider/Rear Rider/Metrics/LocationManager.swift
@@ -1,58 +1,0 @@
-//
-//  LocationManager.swift
-//  Rear Rider
-//
-//  Created by Paul Sutton on 11/30/22.
-//
-
-import Foundation
-import CoreLocation
-
-class LocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
-    @Published var authorizationStatus: CLAuthorizationStatus
-    @Published var lastSeenLocation: CLLocation?
-    @Published var currentPlacemark: CLPlacemark?
-    
-    private let locationManager: CLLocationManager
-    
-    override init() {
-        locationManager = CLLocationManager()
-        authorizationStatus = locationManager.authorizationStatus
-        
-        super.init()
-        locationManager.delegate = self
-        locationManager.desiredAccuracy = kCLLocationAccuracyBestForNavigation
-        locationManager.distanceFilter = kCLDistanceFilterNone
-        locationManager.allowsBackgroundLocationUpdates = true
-        locationManager.showsBackgroundLocationIndicator = true
-    }
-    
-    func requestPermission() {
-        locationManager.requestWhenInUseAuthorization()
-    }
-    
-    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
-        authorizationStatus = manager.authorizationStatus
-    }
-    
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        lastSeenLocation = locations.first
-        fetchCountryAndCity(for: locations.first)
-    }
-    
-    func startTracking() {
-        locationManager.startUpdatingLocation()
-    }
-    
-    func stopTracking() {
-        locationManager.stopUpdatingLocation()
-    }
-    
-    func fetchCountryAndCity(for location: CLLocation?) {
-        guard let location = location else { return }
-        let geocoder = CLGeocoder()
-        geocoder.reverseGeocodeLocation(location) { (placemarks, error) in
-            self.currentPlacemark = placemarks?.first
-        }
-    }
-}
