+++ Rear Rider/Rear Rider/Metrics/TrackingManager.swift
@@ -1,13 +1,36 @@
 //
-//  TrackingManager.swift
+//  LocationManager.swift
 //  Rear Rider
 //
-//  Created by Paul Sutton on 12/1/22.
+//  Created by Paul Sutton on 11/30/22.
 //
 
 import Foundation
+import CoreLocation
 
-class TrackingManger: ObservableObject {
+class TrackingManager: NSObject, ObservableObject, CLLocationManagerDelegate {
+    @Published var authorizationStatus: CLAuthorizationStatus
+    @Published var lastSeenLocation: CLLocation?
+    @Published var currentPlacemark: CLPlacemark?
+    @Published var secondsElapsed = 0.0
+    @Published var mode: TrackingMode = .stopped
+    
+    private var timer = Timer()
+    
+    private let locationManager: CLLocationManager
+    
+    override init() {
+        locationManager = CLLocationManager()
+        authorizationStatus = locationManager.authorizationStatus
+        
+        super.init()
+        locationManager.delegate = self
+        locationManager.desiredAccuracy = kCLLocationAccuracyBestForNavigation
+        locationManager.distanceFilter = kCLDistanceFilterNone
+        locationManager.allowsBackgroundLocationUpdates = true
+        locationManager.showsBackgroundLocationIndicator = true
+    }
+    
     
     enum TrackingMode {
         case started
@@ -15,31 +38,45 @@ class TrackingManger: ObservableObject {
         case paused
     }
     
-    @Published var secondsElapsed = 0.0
-    @Published var mode: TrackingMode = .stopped
+    func requestPermission() {
+        locationManager.requestWhenInUseAuthorization()
+    }
     
-    var timer = Timer()
-    var locationManger = LocationManager()
+    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
+        authorizationStatus = manager.authorizationStatus
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        lastSeenLocation = locations.first
+        fetchCountryAndCity(for: locations.first)
+    }
+    
+    func fetchCountryAndCity(for location: CLLocation?) {
+        guard let location = location else { return }
+        let geocoder = CLGeocoder()
+        geocoder.reverseGeocodeLocation(location) { (placemarks, error) in
+            self.currentPlacemark = placemarks?.first
+        }
+    }
     
     func start() {
-        locationManger.startTracking()
         timer = Timer.scheduledTimer(withTimeInterval: 0.1, repeats: true) { timer in
             self.secondsElapsed = self.secondsElapsed + 0.1
         }
+        locationManager.startUpdatingLocation()
         mode = .started
     }
     
     func pause() {
-        locationManger.stopTracking()
+        locationManager.stopUpdatingLocation()
         timer.invalidate()
         mode = .paused
     }
         
     func stop() {
-        locationManger.stopTracking()
+        locationManager.stopUpdatingLocation()
         timer.invalidate()
         secondsElapsed = 0
         mode = .stopped
     }
-    
 }
